split out env
