+++ cross/Cargo.toml
@@ -32,7 +32,11 @@ embassy-usb = { version = "0.4.0", features = ["defmt", "max-handler-count-6"] }
 
 defmt-rtt = { version = "0.4" }
 defmt = { version = "0.3" }
-
+cortex-m = { version = "0.7.7", features = [
+  "inline-asm",
+  "critical-section-single-core",
+] }
+cortex-m-rt = "0.7"
 lib = { path = "../lib" }
 static_cell = { version = "2.1.0" }
