+++ cross/rust-toolchain.toml
@@ -1,0 +1,4 @@
+[toolchain]
+channel = "nightly"
+components = ["rust-src", "rustfmt"]
+targets = ["thumbv7em-none-eabi"]
